[DOC][REFACTOR] Migrate get_history and Fix quickstart (#29)
* Migrate get_history from utils to __init__
* fix README & quickstart.rst
